backend/requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
langchain==0.3.14
langchain_groq==0.2.4
langchain_community==0.3.14
langchain_huggingface==0.1.2
faiss-cpu==1.9.0.post1
pypdf==5.1.0
python-multipart==0.0.20
sentence-transformers==3.3.1
pydantic==2.10.4
python-dotenv==1.0.1
